Converted to use Widget Factory (Untested)
After discussing on IRC how to best structure the plugin, I decided to
opt for leveraging the new jQueryUI v1.9 Widget Factory.

I realize it's not commonly used outside of UI but I figure it would be
good to lead by example for proper plugin development.

This conversion is what I call "theoretical programming" and therefore
has yet to be tested.